refactor(ecs): rewrite configuration system.
